Be a bit more exacting with error messages when parsing numbers.
In particular, the lexer has no idea about the range of numbers,
so we shouldn't claim that there is a bug in the program if we
can't handle an integer that can't fit into 64 bits. Similarly
with floats.

+++ decode.go
@@ -194,7 +194,7 @@ func unifyFloat64(data interface{}, rv reflect.Value) error {
 }
 
 func unifyInt(data interface{}, rv reflect.Value) error {
-	if num, ok := data.(int); ok {
+	if num, ok := data.(int64); ok {
 		rv.SetInt(int64(num))
 		return nil
 	}
